Update process_param.f90
The change takes care of the case if the shape function and time scale are less than the modelling time step. if the UH are less than the modelling time step they set the UH length to 1. This was creating problems for the previous implementation. I have tested it with my set up; for daily it set UH length to 1 and for hourly simulation it set it to 4 or 5 hours which is actually what is should do...