Misc changes to get the mosart test case working
Use the reordered ntopo file for mosart.

This gets the mosart test case roughly working for
a low PE count (as long as the right ESMF mesh is used).

Get ix1, ix2 and indices correct for different task counts.

Change some references from mosart to mizuroute. Add some
diagnostic info. on the ESMF DistGrid and Mesh that's created.
Change debug variable to debug_write. Also change the ESMF_Mesh
create into one call rather than two, like the latest MOSART.
Add some shr_sys_flush calls so output is flushed after being
done.

Some changes that get it closer for the single task MPI case to
work. Check if mpi is initialized and if it is, keep going for
some cases when it would return.

Make sure volr and flood is just set to zero, as these
are unneeded now.

Also added a quick-sort for intel for gindex which turns out is
not needed.